requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database (Async)
sqlmodel==0.0.16
asyncpg==0.29.0
greenlet==3.0.3
alembic==1.13.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
python-dotenv==1.0.1

# Cache & Rate Limiting
redis==5.0.3
slowapi==0.1.9

# Validation & Settings
pydantic-settings==2.2.1
email-validator==2.1.0.post1